+++ decode.go
@@ -1,7 +1,6 @@
 package pq
 import "encoding/hex"
 import "strconv"
-import "time"
 
 func decode(s []byte, typ oid) interface{} {
 	switch typ {
@@ -44,26 +43,5 @@ func decode(s []byte, typ oid) interface{} {
 	}
 
 	return s
-}
-
-func mustParse(f string, typ oid, s []byte) time.Time {
-	str := string(s)
-
-	// Special case until time.Parse bug is fixed:
-	// http://code.google.com/p/go/issues/detail?id=3487
-	if str[len(str)-2] == '.' {
-		str += "0"
-	}
-
-	// check for a 30-minute-offset timezone
-	if (typ == t_timestamptz || typ == t_timetz) &&
-		str[len(str)-3] == ':' {
-		f += ":00"
-	}
-	t, err := time.Parse(f, str)
-	if err != nil {
-		errorf("decode: %s", err)
-	}
-	return t
 }
